ci: move addons-bundle task to the build kind
It turns out that Chain of Trust has a hardcoded list of kinds that are
allowed to be part of a release pipeline and fails due to this being an
'addons' task.

We could update this list, but that would involve a new scriptworker
release + rollout. It's simpler to move this over to the build kind
instead.

DO NOT LAND: add a dummy task which uploads a release artifact